Rework test structure (WIP)
