+++ tests/devices/default.test.js
@@ -1,0 +1,73 @@
+const Device = require('../../functions/devices/default.js');
+
+describe('Test Fan Device', () => {
+  const item = {
+    "type": "Number",
+    "state": "50",
+    "name": "DefaultDevice",
+    "label": "Default Device",
+    "metadata": {
+      "ga": {
+        "value": "",
+        "config": {
+          "inverted": true,
+          "ackNeeded": true,
+          "pinNeeded": "1234"
+        }
+      }
+    }
+  };
+
+  test('Test getConfig', async () => {
+    expect(Device.getConfig(item)).toStrictEqual({
+      "ackNeeded": true,
+      "inverted": true,
+      "pinNeeded": "1234"
+    });
+  });
+
+  test('Test getState', async () => {
+    expect(Device.getState(item)).toStrictEqual({});
+  });
+
+  test('Test getMetadata', async () => {
+    expect(Device.getMetadata(item)).toStrictEqual({
+      "attributes": {},
+      "customData": {
+        "ackNeeded": true,
+        "deviceType": "DefaultDevice",
+        "inverted": true,
+        "itemType": "Number",
+        "pinNeeded": "1234"
+      },
+      "deviceInfo": {
+        "hwVersion": "2.5.0",
+        "manufacturer": "openHAB",
+        "model": "Number:DefaultDevice",
+        "swVersion": "2.5.0",
+      },
+      "id": "DefaultDevice",
+      "name": {
+        "defaultNames": [
+          "Default Device",
+        ],
+        "name": "Default Device",
+        "nicknames": [
+          "Default Device",
+        ],
+      },
+      "roomHint": undefined,
+      "structureHint": undefined,
+      "traits": [],
+      "type": "",
+      "willReportState": false
+    });
+  });
+
+  test('Test hasTag', async () => {
+    expect(Device.hasTag({}, "testtag")).toBe(false);
+    expect(Device.hasTag({ "tags": ["test"] }, "testtag")).toBe(false);
+    expect(Device.hasTag({ "tags": ["testtag"] }, "testtag")).toBe(true);
+    expect(Device.hasTag({ "tags": ["TestTag"] }, "testtag")).toBe(true);
+  });
+});
